+++ src/render/glyph_manager.ts
@@ -27,17 +27,15 @@ export class GlyphManager {
     localIdeographFontFamily: string | false;
     entries: {[stack: string]: Entry};
     url: string;
-    language?: string;
 
     // exposed as statics to enable stubbing in unit tests
     static loadGlyphRange = loadGlyphRange;
     static TinySDF = TinySDF;
 
-    constructor(requestManager: RequestManager, localIdeographFontFamily?: string | false, language?: string) {
+    constructor(requestManager: RequestManager, localIdeographFontFamily?: string | false) {
         this.requestManager = requestManager;
         this.localIdeographFontFamily = localIdeographFontFamily;
         this.entries = {};
-        this.language = language;
     }
 
     setURL(url?: string | null) {
@@ -171,7 +169,7 @@ export class GlyphManager {
                 buffer: 3 * textureScale,
                 radius: 8 * textureScale,
                 cutoff: 0.25,
-                lang: this.language,
+                lang: document?.documentElement?.lang,
                 fontFamily,
                 fontWeight
             });